Fixing of file structure for Unit and Integration Tests with example of UnitTest -Msi
