+++ nest-back/test/Workout/IntegrationTests/createExerciseIntegration.spec.ts
@@ -1,7 +1,6 @@
-import { MockContext, Context, createMockContext } from '../context'
-import {WorkoutService} from "../src/Workout/workout.service";
-import {WorkoutController} from "../src/Workout/workout.controller";
-import { Test, TestingModule } from '@nestjs/testing';
+import { MockContext, Context, createMockContext } from "../../../context";
+import {WorkoutService} from "../../../src/Workout/workout.service";
+//import { Test, TestingModule } from '@nestjs/testing';
 import {v4 as uuidv4 } from 'uuid';
 import {
     Workout,
@@ -10,10 +9,6 @@ import {
     Difficulty,
     Prisma
 } from '@prisma/client';
-import {randomUUID} from "crypto";
-import {WorkoutModule} from "../src/Workout/workout.module";
-import {executionAsyncResource} from "async_hooks";
-import {PrismaService} from "../src/Prisma/prisma.service";
 
 let mockCtx: MockContext
 let ctx: Context
@@ -25,20 +20,7 @@ beforeEach(() => {
     ctx = (mockCtx as unknown) as Context
 })
 
-test('should create new workout ', async () => {
-    const Workout = {
-        workoutID: uuidv4(),
-        workoutTitle: "postmanTest",
-        workoutDescription: "postmanMoreTest",
-        difficulty: Difficulty.EASY,
-        planner_Email: null
-    }
-    mockCtx.prisma.workout.create.mockResolvedValue(Workout);
 
-    await expect(workoutService.createWorkout(Workout.workoutTitle,Workout.workoutDescription, Workout.difficulty,ctx)).resolves.toEqual(
-        "Workout Created."
-    )
-})
 
 test('should create new exercise ', async () => {
     const Exercise = {
@@ -60,10 +42,3 @@ test('should create new exercise ', async () => {
         "Exercise created."
     )
 })
-
-
-
-
-
-
-
